user-service/requirements.txt
asgiref
Django
django-annoying==0.10.6
djangorestframework
psycopg2
psycopg2-binary>=2.8
pytz==2021.3
six==1.16.0
sqlparse==0.2.4
python-dotenv
gunicorn
requests
redis
django-cors-headers
pika
pyyaml
uritemplate
coreapi
django-cors-headers